mingw: remove the Makefile target to install test artifacts
It is still one of the really vexing quirks of Git's test suite that it
requires executables that are not installed via the `install` target. To
be able to validate MinGit, I introduced a Makefile target once upon a
time to address that omission so as to be able to run the test suite on
any installed Git version by bundling the test suite together with those
test-supporting executables.

However, I never got around to upstream this because the principal
reason (moving toward a BusyBox-backed MinGit) went away: BusyBox-w32 is
simply not maintained in a confidence-inducing manner and therefore I
all but abandoned that initiative.

Therefore, this Makefile target is not in a good shape, and it's time to
let it go.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>